Moved PPP module and associated files into its own sub-package.
